get adversarial training loop running

+++ train_ms.py
@@ -292,11 +292,11 @@ def main():
             _, batch_t = targetloader_iter.__next__()
 
             images, labels, _, _ = batch
-            images = images.cuda()
-            labels = labels.long().cuda()
+            images = images.float().cuda() # images.cuda()
+            labels = labels.long().cuda() # labels.double().cuda()
             images_t, labels_t, _, _ = batch_t
-            images_t = images_t.cuda()
-            labels_t = labels_t.long().cuda()
+            images_t = images_t.float().cuda() # images.cuda()
+            labels_t = labels_t.long().cuda() # labels_t.double().cuda()
 
             with Timer("Elapsed time in update: %f"):
                 loss_seg1, loss_seg2, loss_adv_target1, loss_adv_target2, loss_me, loss_kl, pred1, pred2, pred_target1, pred_target2, val_loss = Trainer.gen_update(images, images_t, labels, labels_t, i_iter)
